+++ browser/src/Services/Configuration/Config.ts
@@ -3,124 +3,17 @@ import * as cloneDeep from "lodash/cloneDeep"
 import * as isError from "lodash/isError"
 import * as path from "path"
 
-import * as Log from "./Log"
-import * as Performance from "./Performance"
-import * as Platform from "./Platform"
+import * as Log from "./../../Log"
+import * as Performance from "./../../Performance"
+import * as Platform from "./../../Platform"
 
-import { Event, IEvent } from "./Event"
+import { Event, IEvent } from "./../Event"
 
-import { applyDefaultKeyBindings } from "./Input/KeyBindings"
+import { applyDefaultKeyBindings } from "./../../Input/KeyBindings"
 
-import { diff } from "./Utility"
+import { diff } from "./../../Utility"
 
-export interface IConfigValues {
-
-    "activate": (oni: Oni.Plugin.Api) => void
-    "deactivate": () => void
-
-    // Debug settings
-    "debug.fixedSize": {
-        rows: number,
-        columns: number,
-    } | null
-
-    // Option to override neovim path. Used for testing new versions before bringing them in.
-    "debug.neovimPath": string | null
-
-    // Production settings
-
-    // Bell sound effect to use
-    // See `:help bell` for instances where the bell sound would be used
-    "oni.audio.bellUrl": string
-
-    // The default config is an opinionated, prescribed set of plugins. This is on by default to provide
-    // a good out-of-box experience, but will likely conflict with a Vim/Neovim veteran's finely honed config.
-    //
-    // Set this to 'false' to avoid loading the default config, and load settings from init.vim instead.
-    "oni.useDefaultConfig": boolean
-
-    // By default, user's init.vim is not loaded, to avoid conflicts.
-    // Set this to `true` to enable loading of init.vim.
-    "oni.loadInitVim": boolean
-
-    // Sets the `popupmenu_external` option in Neovim
-    // This will override the default UI to show a consistent popupmenu,
-    // whether using Oni's completion mechanisms or VIMs
-    //
-    // Use caution when changing the `menuopt` parameters if using
-    // a custom init.vim, as that may cause problematic behavior
-    "oni.useExternalPopupMenu": boolean
-
-    // If true, hide Menu bar by default
-    // (can still be activated by pressing 'Alt')
-    "oni.hideMenu": boolean
-
-    // glob pattern of files to exclude from fuzzy finder (Ctrl-P)
-    "oni.exclude": string[]
-
-    // bookmarks to open if opened in install dir
-    "oni.bookmarks": string[]
-
-    // Editor settings
-
-    "editor.backgroundOpacity": number
-    "editor.backgroundImageUrl": string
-    "editor.backgroundImageSize": string
-
-    // Setting this to true enables yank integration with Oni
-    // When true, and text is yanked / deleted, that text will
-    // automatically be put on the clipboard.
-    //
-    // In addition, this enables <C-v> and <Cmd-v> behavior
-    // in paste from clipboard in insert mode.
-    "editor.clipboard.enabled": boolean
-
-    "editor.quickInfo.enabled": boolean
-    // Delay (in ms) for showing QuickInfo, when the cursor is on a term
-    "editor.quickInfo.delay": number
-
-    "editor.completions.enabled": boolean
-    "editor.errors.slideOnFocus": boolean
-    "editor.formatting.formatOnSwitchToNormalMode": boolean // TODO: Make this setting reliable. If formatting is slow, it will hose edits... not fun
-
-    // If true (default), ligatures are enabled
-    "editor.fontLigatures": boolean
-    "editor.fontSize": string
-    "editor.fontFamily": string // Platform specific
-
-    // Additional padding between lines
-    "editor.linePadding": number
-
-    // If true (default), the buffer scroll bar will be visible
-    "editor.scrollBar.visible": boolean
-
-    // Additional paths to include when launching sub-process from Oni
-    // (and available in terminal integration, later)
-    "environment.additionalPaths": string[]
-
-    // Command to list files for 'quick open'
-    // For example, to use 'ag': ag --nocolor -l .
-    //
-    // The command must emit a list of filenames
-    //
-    // IE, Windows:
-    // "editor.quickOpen.execCommand": "dir /s /b"
-    "editor.quickOpen.execCommand": string | null
-
-    "editor.fullScreenOnStart": boolean
-
-    "editor.cursorLine": boolean
-    "editor.cursorLineOpacity": number
-
-    "editor.cursorColumn": boolean
-    "editor.cursorColumnOpacity": number
-
-    "statusbar.enabled": boolean
-    "statusbar.fontSize": string
-
-    "tabs.enabled": boolean
-    "tabs.showVimTabs": boolean
-}
+import { IConfigurationValues } from "./IConfigurationValues"
 
 const noop = () => { } // tslint:disable-line no-empty
 
@@ -188,7 +81,7 @@ export class Config {
         "tabs.showVimTabs": false,
     }
 
-    private MacConfig: Partial<IConfigValues> = {
+    private MacConfig: Partial<IConfigurationValues> = {
         "editor.fontFamily": "Menlo",
         "environment.additionalPaths": [
             "/usr/bin",
@@ -196,11 +89,11 @@ export class Config {
         ],
     }
 
-    private WindowsConfig: Partial<IConfigValues> = {
+    private WindowsConfig: Partial<IConfigurationValues> = {
         "editor.fontFamily": "Consolas",
     }
 
-    private LinuxConfig: Partial<IConfigValues> = {
+    private LinuxConfig: Partial<IConfigurationValues> = {
         "editor.fontFamily": "DejaVu Sans Mono",
         "environment.additionalPaths": [
             "/usr/bin",
@@ -210,13 +103,13 @@ export class Config {
 
     private DefaultPlatformConfig = Platform.isWindows() ? this.WindowsConfig : Platform.isLinux() ? this.LinuxConfig : this.MacConfig
 
-    private _onConfigurationChangedEvent: Event<Partial<IConfigValues>> = new Event<Partial<IConfigValues>>()
+    private _onConfigurationChangedEvent: Event<Partial<IConfigurationValues>> = new Event<Partial<IConfigurationValues>>()
 
     private _oniApi: Oni.Plugin.Api = null
 
-    private _config: IConfigValues = null
+    private _config: IConfigurationValues = null
 
-    public get onConfigurationChanged(): IEvent<Partial<IConfigValues>> {
+    public get onConfigurationChanged(): IEvent<Partial<IConfigurationValues>> {
         return this._onConfigurationChangedEvent
     }
 
@@ -245,11 +138,11 @@ export class Config {
         Performance.mark("Config.load.end")
     }
 
-    public hasValue(configValue: keyof IConfigValues): boolean {
+    public hasValue(configValue: keyof IConfigurationValues): boolean {
         return !!this.getValue(configValue)
     }
 
-    public getValue<K extends keyof IConfigValues>(configValue: K, defaultValue?: any) {
+    public getValue<K extends keyof IConfigurationValues>(configValue: K, defaultValue?: any) {
         if (typeof this._config[configValue] === "undefined") {
             return defaultValue
         } else {
@@ -257,7 +150,7 @@ export class Config {
         }
     }
 
-    public getValues(): IConfigValues {
+    public getValues(): IConfigurationValues {
         return cloneDeep(this._config)
     }
 
@@ -313,8 +206,8 @@ export class Config {
         }
     }
 
-    private getUserRuntimeConfig(): IConfigValues | Error {
-        let userRuntimeConfig: IConfigValues | null = null
+    private getUserRuntimeConfig(): IConfigurationValues | Error {
+        let userRuntimeConfig: IConfigurationValues | null = null
         let error: Error | null = null
         if (fs.existsSync(this.userJsConfig)) {
             try {
